Add non-ground navigation targets
Add the possibility for a player to click on objects in the level and
have the character move towards these objects. Later on this will allow
players to interact with such objects.

On the side, this commit also fixes a bug in which the character would
not finish turning towards the target if there was not enough time
during its movement. That is, the turn speed was not fast enough, so it
would reach the target before it finished turning towards it. Now, it
will continue to turn, even if it reached the target.