Refactor target desired distance
This diff uses the `target_desired_distance` variable from `NavTarget`
to set a padding between the Player character and its interactable
targets. Targets on the ground will not have this padding applied.

I have later decided to drop this idea and just use collision shapes
to avoid having a player character stand awkwardly close to a target
enemy character. This diff is moved to a separate branch in case I find
another use for this code.